Swapped IncrementalPIDController to PositionalPIDController
After implementing the PID to optimal algo, we found out the PID control is not very stable, and it is not due to the algo itself.
We believe IncrementalPIDController is not stable and safe to use anyway, so we are changing to PositionalPIDController.
The values need to be tuned tmr, and pid_tuning.cpp has been changed to accomodate such change.
Hope this works.